chore: 🐋 Image 컴포넌트에 index 추가
