+++ src/components/atoms/Image/Image.types.ts
@@ -1,9 +1,9 @@
 import { ImageProps } from "next/image";
 import { VariantProps } from "tailwind-variants";
 
-import { NonNullableProps } from "@customTypes/utilType";
+import type { NonNullableProps } from "@customTypes/utilType";
 
-import { imageVariants } from "./Image";
+import { imageVariants } from ".";
 
 /** cva로 만든 Image Variants의 Type */
 type ImageVariantsType = NonNullableProps<VariantProps<typeof imageVariants>>;
