+++ src/components/atoms/Image/index.ts
@@ -1,0 +1,9 @@
+import Image from "./Image";
+import { imageVariants } from "./Image.styles";
+import type { ImagePropsType } from "./Image.types";
+
+export type { ImagePropsType };
+
+export { imageVariants };
+
+export default Image;
